Fix installation of the printer-driver-cnijfilter meta-package on amd64
Many (75) of the packages that are listed as dependencies of this meta-package
are 'Architecture: i386' only, and that is blocking the installation of the
whole suite of drivers on amd64, even if 58 packages in the list would still
be suitable for 64bit systems.

To fix this problem, let's mark the i386-only packages as [i386] in the list of
dependencies for the meta-package so that they won't block the installation of
the whole suite on amd64.

https://phabricator.endlessm.com/T12338