Switch build dependency from libtiff4-dev to libtiff-dev
CUPS 2.2 depends now on libtiff5, so trying to keep libtiff4 as a
build-dep here causes a conflict and breaks the build innecessarily,
as this package should build and work just fine with libtiff5 too.

We do depend on the libtiff-dev metapackage, though, which is provided
since libtiff5 (it was not for libtiff4) and is more future-proof.

https://phabricator.endlessm.com/T15707