Add missing drivers from the 4.00 release
For some reason there are a bunch of drivers from 4.00 that,
while they made it to our debian package in OBS, they have not
been added to this git repository along with commit 0460c7e5
so we need to add them back now or we will lose support for
the following models: MG7100, MG6500, MG6400, MG5500, MG3500,
MG2400, MG2500 and P200.

This commit adds those files back, which should make the packages
built from this git repository identical to the ones previously
built in OBS. Note that no changes are needed to the debian files
since those were already included with commit 0058e363.

https://phabricator.endlessm.com/T16100